requirements.txt
# fastapi web framework
uvicorn[standard] == 0.15.0
gunicorn == 20.1.0
email-validator == 1.0.5
emails == 0.5.15
pydantic >= 1.9, <2.0
fastapi == 0.68.1
python-multipart == 0.0.5
# configurations
hydra-core == 1.2.0
loguru
# pymongo>=4.1 required by motor
pymongo == 4.3.3
motor == 3.0.0
odmantic == 0.9.1

passlib[bcrypt] == 1.7.4
python-jose[cryptography] == 3.1.0
tenacity == 6.1.0
